Moving the Duration wrapper to it's own package to break import loops.
